Fixed bash-isms in entrypoint run by /bin/sh
`[[` is a feature of `bash`, and `entrypoint.sh` has `sh` in it's shebang. So the script dumps warning `[[: not found` somewhere and goes with wrong branch. Same goes for other tests in the script.
IIRC, Alpine doesn't have bash bundled, so simple change of shebang won't help. This commit removes bash-specific syntax and keep with clean POSIX sh.

Fixes #26.